Introduced public API and started developing CLI using it
Added Tabularize command for easy conversion to CSV using CLI.

Added Transform command that replaces a number of Perl scripts used
for sorting, naming, and filtering with the goal to eventually replace
all presently bundled Perl and Python scripts.

The public API should be considered work in progress at this point. At
the time of the stable release, the public API will support a number
of operations and will remain stable over time as the internal program
structure continues to change.